start sparkContext when start worker , resource monitor

+++ grid/src/main/scala/moonbox/grid/deploy/DeployMessage.scala
@@ -15,7 +15,7 @@ object DeployMessages {
 		id: String,
 		worker: ActorRef,
 		cores: Int,
-		memory: Int) extends DeployMessage {
+		memory: Long) extends DeployMessage {
 	}
 
 	case class JobStateChanged(
